+++ src/components/Visualizations/Graphs/index.ts
@@ -1,4 +1,4 @@
 export * from './BarGraph';
 export * from './LineGraph';
-export * from './NetworkGraph';
+export * from './AuthorNetworkGraph';
 export * from './PaperNetworkGraph';
